requirements.txt
watchdog
pillow
dbus-python
black
flake8